+++ scripts/build-tl-components-configuration.js
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-
 const dotEnv = require('dotenv-override-true');
 const fs = require('fs-extra');
 
@@ -26,7 +25,6 @@ if(fs.existsSync('./dist')) {
 
 if(fs.existsSync('./static')) {
   fs.writeFile('./static/config.json', configTemplate);
-  fs.writeFile('./static/tamu-library-components/docs/usage/config.json', configTemplate);
 
   fs.copy('./static/config.json', './static/tamu-library-components/config.json', err => {
     if (err) return console.error('Could not copy config file to static/tamu-library-components', err);
@@ -38,10 +36,16 @@ if(fs.existsSync('./static')) {
     console.log('Copied config file to static/tamu-library-components/docs successfully');
   });
 
+  fs.copy('./static/config.json', './static/tamu-library-components/docs/usage/config.json', err => {
+    if (err) return console.error('Could not copy config file to static/tamu-library-components/docs/usage path', err);
+    console.log('Copied config file to static/tamu-library-components/docs/usage successfully');
+  });
+
   fs.copy('./static/config.json', './static/tamu-library-components/docs/usage/assets/config.json', err => {
     if (err) return console.error('Could not copy config file to static/tamu-library-components/docs/usage/assets path', err);
     console.log('Copied config file to static/tamu-library-components/docs/usage/assets successfully');
   });
+
 }
 
 fs.writeFile('./src/config.json', configTemplate);